Remove validation app
