+++ djangomain/urls.py
@@ -49,14 +49,13 @@ urlpatterns = [
     path("api/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
     path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
     path("api/token/verify/", TokenVerifyView.as_view(), name="token_verify"),
-    path("api/measurement/", include("measurement.urls", namespace="measurement")),
-    path("api/importing/", include("importing.urls", namespace="importing")),
+    path("measurement/", include("measurement.urls", namespace="measurement")),
+    path("importing/", include("importing.urls", namespace="importing")),
     path("django_plotly_dash/", include("django_plotly_dash.urls")),
     # path("station/", include("station.urls", namespace="station")),
     # path("sensor/", include("sensor.urls", namespace="sensor")),
     # path("variable/", include("variable.urls", namespace="variable")),
     # path("formatting/", include("formatting.urls", namespace="formatting")),
-    path("validated/", include("validated.urls", namespace="validated")),
     # re_path(
     #     r"^swagger(?P<format>\.json|\.yaml)$",
     #     schema_view.without_ui(cache_timeout=0),
