+++ measurement/dash_apps/data_report.py
@@ -3,10 +3,11 @@ import plotly.express as px
 from dash import Input, Output, State, dcc, html
 from django_plotly_dash import DjangoDash
 
-from measurement.reporting import get_report_data_from_db
-from validated.filters import get_date_range, get_station_options, get_variable_options
-from validated.plots import create_empty_plot, create_report_plot
 from variable.models import Variable
+
+from ..filters import get_date_range, get_station_options, get_variable_options
+from ..reporting import get_report_data_from_db
+from .plots import create_empty_plot, create_report_plot
 
 # Create a Dash app
 app = DjangoDash(
