+++ measurement/templates/daily_validation.html
