+++ measurement/urls.py
@@ -20,6 +20,8 @@ app_name = "measurement"
 urlpatterns = [
     path("measurement/", views.MeasurementList.as_view()),
     path("report/", views.ReportList.as_view()),
+    path("data_report/", views.DataReport.as_view(), name="data_report"),
+    path("daily_validation/", views.DailyValidation.as_view(), name="daily_validation"),
 ]
 
 urlpatterns = format_suffix_patterns(urlpatterns)
