--- validated/apps.py
@@ -1,20 +1,0 @@
-########################################################################################
-# Plataforma para la Iniciativa Regional de Monitoreo Hidrológico de Ecosistemas Andinos
-# (iMHEA)basada en los desarrollos realizados por:
-#     1) FONDO PARA LA PROTECCIÓN DEL AGUA (FONAG), Ecuador.
-#           Contacto: info@fonag.org.ec
-#     2) EMPRESA PÚBLICA METROPOLITANA DE AGUA POTABLE Y SANEAMIENTO DE QUITO (EPMAPS),
-#           Ecuador.
-#           Contacto: paramh2o@aguaquito.gob.ec
-#
-#  IMPORTANTE: Mantener o incluir esta cabecera con la mención de las instituciones
-#  creadoras, ya sea en uso total o parcial del código.
-########################################################################################
-
-from __future__ import unicode_literals
-
-from django.apps import AppConfig
-
-
-class ValidatedConfig(AppConfig):
-    name = "validated"
