--- validated/scripts/tests.py
@@ -1,6 +1,0 @@
-from django.test import TestCase
-
-# Create your tests here.
-from val2.functions import reporte_subhorario
-
-reporte_subhorario(1, 2, 3, 4, 5, 30)
