--- validated/static/validated/style.css
@@ -1,51 +1,0 @@
-.normal {background-color: rgba(0, 255, 0, 0.15) !important;}
-.error {background-color: rgba(255, 0, 0, 0.15) !important;}
-.warning {background-color: rgba(255, 255, 0, 0.15)  !important;}
-
-.nopadding {
-    padding: 0;
-    margin: 0;
-}
-
-.right-panel {
-    height: calc(100vh - 150px);
-    overflow: hidden;
-}
-
-.validation-body {
-    overflow-y: hidden;
-}
-
-.report-body {
-    overflow-y: hidden;
-}
-
-
-.popup-overlay {
-    visibility:hidden;
-    position:fixed;
-    background-size:100% 100%;
-    height: 100%;
-    width: 100% ;
-    left: 0;
-    top: 0;
-    z-index: 1;
-    background-color: rgb(0,0,0);
-    background-color: rgba(0,0,0,0.4);
-}
-
-.popup-content {
-    position:relative;
-    margin:0 auto;
-    visibility:hidden;
-    height:20%;
-    width:40%;
-    vertical-align: middle;
-}
-.message{
-    vertical-align: middle;
-    text-align: center;
-}
-.popup-overlay.active{visibility:visible;}
-.popup-content.active{visibility:visible;}
-
