--- validated/templates/daily_validation_right_panel.html
@@ -1,23 +1,0 @@
-<ul class="nav nav-tabs" id="myTabs" role="tablist">
-  <li class="nav-item">
-    <a class="nav-link active" onclick="getTab(event, 'tab1')" id="tab1-tab" data-toggle="tab" href="#tab1" role="tab" aria-controls="tab1" aria-selected="true">DAILY REPORT</a>
-  </li>
-  <li class="nav-item ">
-    <a class="nav-link" onclick="getTab(event, 'tab2')" id="tab2-tab" data-toggle="tab" href="#tab2" role="tab" aria-controls="tab2" aria-selected="false">PLOT</a>
-  </li>
-  <li class="nav-item" >
-    <a class="nav-link" style="display:none;" onclick="getTab(event, 'tab3')" id="tab3-tab" data-toggle="tab" href="#tab3" role="tab" aria-controls="tab3" aria-selected="false">DETAIL OF SELECTED DAY</a>
-  </li>
-</ul>
-
-<div class="tab-content" id="myTabsContent">
-  <div class="tab-pane fade show active" id="tab1" role="tabpanel" aria-labelledby="tab1-tab">
-        {% include "daily_validation_tab1.html" %}
-  </div>
-  <div class="tab-pane fade" id="tab2" role="tabpanel" aria-labelledby="tab2-tab">
-      {% include "daily_validation_tab2.html" %}
-  </div>
-  <div class="tab-pane fade" id="tab3" role="tabpanel" aria-labelledby="tab3-tab">
-      {% include "daily_validation_tab3.html" %}
-  </div>
-</div>
