--- validated/templates/daily_validation_tab1.html
@@ -1,20 +1,0 @@
-    <div id="toolbar_daily" class="filter row no-gutters">
-        <div class="col-lg-10 col-md-10 col-sm-10">
-        </div>
-
-        <div class="col-lg-2 col-md-2 col-sm-2">
-            <div class="btn-group">
-                <button id="btn_daily_send" name="daily" class="btn btn-xs btn-success" data-toggle="tooltip" data-placement="top" title="Send validation">Save to validated</i></button>
-            </div>
-        </div>
-    </div>
-
-
-    <table id="table_daily"
-        data-toggle="table"
-        data-height="250"
-        data-show-footer="true"
-        data-unique-id="id"
-        data-show-columns="true"
-        data-toolbar="#toolbar_daily" class="filter">
-    </table>
