--- validated/templates/daily_validation_tab2.html
@@ -1,2 +1,0 @@
-    <div id="div_information" style="overflow-x: scroll;"  class="col-lg-12 col-md-12 col-sm-12 nopadding">
-    </div>
