--- validated/templates/daily_validation_tab3.html
@@ -1,82 +1,0 @@
-<input type="hidden" id="orig_detail_date">
-<div class=" row my-3">
-    <div class="col-lg-2 col-md-2 col-sm-6">
-        <b>Selected filter</b>
-        <select id="chk_detail_selected" name="detail" type="text" class="form-control mr-2">
-            <option value="all">All</option>
-            <option value="selected">Selected</option>
-            <option value="non-selected">Non selected</option>
-
-        </select>
-    </div>
-    <div class="col-lg-2 col-md-2 col-sm-6">
-        <b>Time filter</b>
-        <select id="chk_detail_time" name="detail" type="text" class="form-control mr-2">
-            <option value="all">All</option>
-            <option value="shorter_than_tx_period">&lt; Tx Period</option>
-            <option value="greater_than_tx_period">&gt; Tx Period</option>
-        </select>
-    </div>
-    <div class="col-lg-2 col-md-2 col-sm-6">
-        <b>Value filter</b>
-        <select id="chk_detail_value" name="detail" type="text" class="form-control mr-2">
-            <option value="all">All</option>
-            <option value="normal">Below threshold</option>
-            <option value="error">Above threshold</option>
-
-        </select>
-    </div>
-    <div class="col-lg-2 col-md-2 col-sm-6">
-        <b>Outliers filter</b>
-        <select id="chk_detail_stddev" name="detail" type="text" class="form-control mr-2">
-            <option value="all">All</option>
-            <option value="normal">No outliers</option>
-            <option value="error">Outliers</option>
-
-        </select>
-    </div>
-    <div class="col-lg-2 col-md-2 col-sm-6">
-        <b>Value diff. filter</b>
-        <select id="chk_detail_value_difference" name="detail" type="text" class="form-control mr-2">
-            <option value="all">All</option>
-            <option value="normal">Normal</option>
-            <option value="error">Error</option>
-        </select>
-    </div>
-
-</div>
-
-
-<div id="toolbar_detail" class="filter row no-gutters">
-    <div class="col-lg-3 col-md-3 col-sm-3">
-        <input id="txt_detail_selection" type="text" class="form-control" placeholder="1,5,7 or 1-10">
-    </div>
-    <div class="col-lg-5 col-md-5 col-sm-5">
-        <div class="btn-group ml-lg-2 ml-md-2">
-            <button id="btn_detail_select" name="detail" class="btn btn-warning btn-xs" data-toggle="tooltip" data-placement="top" title="Select rows">Select rows</button>
-            <button id="btn_detail_unselect" name="detail" class="btn btn-warning btn-xs" data-toggle="tooltip" data-placement="top" title="Unselect rows">Unselect rows</button>
-        </div>
-    </div>
-    <div class="col-lg-2 col-md-2 col-sm-2">
-        <div class="btn-group">
-            <button id="btn_detail_new" name="detail" class="btn btn-primary btn-xs" data-toggle="tooltip" data-placement="top" title="New row">New</button>
-        </div>
-    </div>
-    <div class="col-lg-2 col-md-2 col-sm-2">
-        <div class="btn-group">
-            <button id="btn_detail_save" name="detail" class="btn btn-success btn-xs" data-toggle="tooltip" data-placement="top" title="Save changes">Save day to Validated</button>
-        </div>
-    </div>
-</div>
-
-
-<table id="table_detail"
-       data-maintain-meta-data="true"
-        data-toggle="table"
-        data-height="380"
-        data-show-footer="true"
-        data-unique-id="id"
-        data-show-columns="true"
-        data-toolbar="#toolbar_detail"
-        class="filter table-sm">
-</table>
